+++ src/RcppExports.cpp
@@ -7,7 +7,7 @@ using namespace Rcpp;
 
 // dgamma_c
 NumericVector dgamma_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_dgamma_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_dgamma_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -19,7 +19,7 @@ END_RCPP
 }
 // deriv_dgamma_c
 NumericVector deriv_dgamma_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_deriv_dgamma_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_dgamma_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -31,7 +31,7 @@ END_RCPP
 }
 // lt_dgamma_c
 double lt_dgamma_c(int m, double s, double theta);
-RcppExport SEXP frailtyr_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -44,7 +44,7 @@ END_RCPP
 }
 // deriv_lt_dgamma_c
 double deriv_lt_dgamma_c(int m, double s, double theta);
-RcppExport SEXP frailtyr_deriv_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -57,7 +57,7 @@ END_RCPP
 }
 // deriv_deriv_lt_dgamma_c
 double deriv_deriv_lt_dgamma_c(int m, double s, double theta);
-RcppExport SEXP frailtyr_deriv_deriv_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_deriv_lt_dgamma_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -70,7 +70,7 @@ END_RCPP
 }
 // lt_dlognormal_c
 double lt_dlognormal_c(int m, double s, double theta);
-RcppExport SEXP frailtyr_lt_dlognormal_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_lt_dlognormal_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -83,7 +83,7 @@ END_RCPP
 }
 // dlognormal_c
 NumericVector dlognormal_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -95,7 +95,7 @@ END_RCPP
 }
 // deriv_dlognormal_c
 NumericVector deriv_dlognormal_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_deriv_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -107,7 +107,7 @@ END_RCPP
 }
 // deriv_deriv_dlognormal_c
 NumericVector deriv_deriv_dlognormal_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_deriv_deriv_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_deriv_dlognormal_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -119,7 +119,7 @@ END_RCPP
 }
 // lt_dinvgauss_c
 double lt_dinvgauss_c(int m, double s, double theta);
-RcppExport SEXP frailtyr_lt_dinvgauss_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_lt_dinvgauss_c(SEXP mSEXP, SEXP sSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -132,7 +132,7 @@ END_RCPP
 }
 // dinvgauss_c
 NumericVector dinvgauss_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -144,7 +144,7 @@ END_RCPP
 }
 // deriv_dinvgauss_c
 NumericVector deriv_dinvgauss_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_deriv_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -156,7 +156,7 @@ END_RCPP
 }
 // deriv_deriv_dinvgauss_c
 NumericVector deriv_deriv_dinvgauss_c(NumericVector x, NumericVector theta);
-RcppExport SEXP frailtyr_deriv_deriv_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
+RcppExport SEXP frailtySurv_deriv_deriv_dinvgauss_c(SEXP xSEXP, SEXP thetaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -168,7 +168,7 @@ END_RCPP
 }
 // lt_dpvf_coef
 double lt_dpvf_coef(int p, int j, double alpha);
-RcppExport SEXP frailtyr_lt_dpvf_coef(SEXP pSEXP, SEXP jSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_lt_dpvf_coef(SEXP pSEXP, SEXP jSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -181,7 +181,7 @@ END_RCPP
 }
 // deriv_lt_dpvf_coef
 double deriv_lt_dpvf_coef(int p, int j, double alpha);
-RcppExport SEXP frailtyr_deriv_lt_dpvf_coef(SEXP pSEXP, SEXP jSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_deriv_lt_dpvf_coef(SEXP pSEXP, SEXP jSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -194,7 +194,7 @@ END_RCPP
 }
 // deriv_deriv_lt_dpvf_coef
 double deriv_deriv_lt_dpvf_coef(int m, int j, double alpha);
-RcppExport SEXP frailtyr_deriv_deriv_lt_dpvf_coef(SEXP mSEXP, SEXP jSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_deriv_deriv_lt_dpvf_coef(SEXP mSEXP, SEXP jSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -207,7 +207,7 @@ END_RCPP
 }
 // dpvf_c
 NumericVector dpvf_c(NumericVector x, NumericVector alpha);
-RcppExport SEXP frailtyr_dpvf_c(SEXP xSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_dpvf_c(SEXP xSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -219,7 +219,7 @@ END_RCPP
 }
 // lt_dpvf_c
 double lt_dpvf_c(int m, double s, double alpha);
-RcppExport SEXP frailtyr_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -232,7 +232,7 @@ END_RCPP
 }
 // deriv_lt_dpvf_c
 double deriv_lt_dpvf_c(int m, double s, double alpha);
-RcppExport SEXP frailtyr_deriv_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_deriv_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -245,7 +245,7 @@ END_RCPP
 }
 // deriv_deriv_lt_dpvf_c
 double deriv_deriv_lt_dpvf_c(int m, double s, double alpha);
-RcppExport SEXP frailtyr_deriv_deriv_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_deriv_deriv_lt_dpvf_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -258,7 +258,7 @@ END_RCPP
 }
 // dposstab_c
 NumericVector dposstab_c(NumericVector x, NumericVector alpha);
-RcppExport SEXP frailtyr_dposstab_c(SEXP xSEXP, SEXP alphaSEXP) {
+RcppExport SEXP frailtySurv_dposstab_c(SEXP xSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -269,21 +269,21 @@ BEGIN_RCPP
 END_RCPP
 }
 // lt_dposstab_c
-double lt_dposstab_c(int p, double s, double alpha);
-RcppExport SEXP frailtyr_lt_dposstab_c(SEXP pSEXP, SEXP sSEXP, SEXP alphaSEXP) {
+double lt_dposstab_c(int m, double s, double alpha);
+RcppExport SEXP frailtySurv_lt_dposstab_c(SEXP mSEXP, SEXP sSEXP, SEXP alphaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
-    Rcpp::traits::input_parameter< int >::type p(pSEXP);
+    Rcpp::traits::input_parameter< int >::type m(mSEXP);
     Rcpp::traits::input_parameter< double >::type s(sSEXP);
     Rcpp::traits::input_parameter< double >::type alpha(alphaSEXP);
-    __result = Rcpp::wrap(lt_dposstab_c(p, s, alpha));
+    __result = Rcpp::wrap(lt_dposstab_c(m, s, alpha));
     return __result;
 END_RCPP
 }
 // zeta
 double zeta(double s);
-RcppExport SEXP frailtyr_zeta(SEXP sSEXP) {
+RcppExport SEXP frailtySurv_zeta(SEXP sSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -294,7 +294,7 @@ END_RCPP
 }
 // phi_c
 double phi_c(int k, int N_dot, double H_dot, double theta, String frailty, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -312,7 +312,7 @@ END_RCPP
 }
 // phi_prime_c
 double phi_prime_c(int k, int N_dot, double H_dot, double theta, String frailty, int deriv_idx, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_prime_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP deriv_idxSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_prime_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP deriv_idxSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -331,7 +331,7 @@ END_RCPP
 }
 // phi_prime_prime_c
 double phi_prime_prime_c(int k, int N_dot, double H_dot, double theta, String frailty, int deriv_idx_1, int deriv_idx_2, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_prime_prime_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP deriv_idx_1SEXP, SEXP deriv_idx_2SEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_prime_prime_c(SEXP kSEXP, SEXP N_dotSEXP, SEXP H_dotSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP deriv_idx_1SEXP, SEXP deriv_idx_2SEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -351,7 +351,7 @@ END_RCPP
 }
 // bh
 List bh(NumericVector d_, List R_star, List K_, List Y_, List N_, List N_dot, NumericVector beta, NumericVector theta, String frailty, NumericVector weights, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_bh(SEXP d_SEXP, SEXP R_starSEXP, SEXP K_SEXP, SEXP Y_SEXP, SEXP N_SEXP, SEXP N_dotSEXP, SEXP betaSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP weightsSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_bh(SEXP d_SEXP, SEXP R_starSEXP, SEXP K_SEXP, SEXP Y_SEXP, SEXP N_SEXP, SEXP N_dotSEXP, SEXP betaSEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP weightsSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -374,7 +374,7 @@ END_RCPP
 }
 // loglikelihood
 NumericVector loglikelihood(List X_, List K_, List I_, List phi_1_, NumericVector lambda, NumericVector beta);
-RcppExport SEXP frailtyr_loglikelihood(SEXP X_SEXP, SEXP K_SEXP, SEXP I_SEXP, SEXP phi_1_SEXP, SEXP lambdaSEXP, SEXP betaSEXP) {
+RcppExport SEXP frailtySurv_loglikelihood(SEXP X_SEXP, SEXP K_SEXP, SEXP I_SEXP, SEXP phi_1_SEXP, SEXP lambdaSEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -390,7 +390,7 @@ END_RCPP
 }
 // xi_beta
 NumericVector xi_beta(List X_, List I_, List H_, List psi_, int r);
-RcppExport SEXP frailtyr_xi_beta(SEXP X_SEXP, SEXP I_SEXP, SEXP H_SEXP, SEXP psi_SEXP, SEXP rSEXP) {
+RcppExport SEXP frailtySurv_xi_beta(SEXP X_SEXP, SEXP I_SEXP, SEXP H_SEXP, SEXP psi_SEXP, SEXP rSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -405,7 +405,7 @@ END_RCPP
 }
 // xi_theta
 NumericVector xi_theta(List phi_1_, List phi_prime_1_, int r);
-RcppExport SEXP frailtyr_xi_theta(SEXP phi_1_SEXP, SEXP phi_prime_1_SEXP, SEXP rSEXP) {
+RcppExport SEXP frailtySurv_xi_theta(SEXP phi_1_SEXP, SEXP phi_prime_1_SEXP, SEXP rSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -418,7 +418,7 @@ END_RCPP
 }
 // phi_k
 List phi_k(int s, List N_dot_, List H_dot_, NumericVector theta, String frailty, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_k(SEXP sSEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_k(SEXP sSEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -436,7 +436,7 @@ END_RCPP
 }
 // phi_prime_k
 List phi_prime_k(int s, int theta_idx, List N_dot_, List H_dot_, NumericVector theta, String frailty, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_prime_k(SEXP sSEXP, SEXP theta_idxSEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_prime_k(SEXP sSEXP, SEXP theta_idxSEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -455,7 +455,7 @@ END_RCPP
 }
 // phi_prime_prime_k
 List phi_prime_prime_k(int s, int theta_idx_1, int theta_idx_2, List N_dot_, List H_dot_, NumericVector theta, String frailty, int kstart, double abstol, double reltol, int maxit);
-RcppExport SEXP frailtyr_phi_prime_prime_k(SEXP sSEXP, SEXP theta_idx_1SEXP, SEXP theta_idx_2SEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP kstartSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
+RcppExport SEXP frailtySurv_phi_prime_prime_k(SEXP sSEXP, SEXP theta_idx_1SEXP, SEXP theta_idx_2SEXP, SEXP N_dot_SEXP, SEXP H_dot_SEXP, SEXP thetaSEXP, SEXP frailtySEXP, SEXP kstartSEXP, SEXP abstolSEXP, SEXP reltolSEXP, SEXP maxitSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -476,7 +476,7 @@ END_RCPP
 }
 // dH_dbeta
 List dH_dbeta(int s, NumericVector d_, List X_, List K_, List R_, List R_dot_, List R_star, List phi_1_, List phi_2_, List phi_3_, NumericVector Lambda, NumericVector lambda, NumericVector beta, NumericVector theta, String frailty);
-RcppExport SEXP frailtyr_dH_dbeta(SEXP sSEXP, SEXP d_SEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP R_SEXP, SEXP R_dot_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP LambdaSEXP, SEXP lambdaSEXP, SEXP betaSEXP, SEXP thetaSEXP, SEXP frailtySEXP) {
+RcppExport SEXP frailtySurv_dH_dbeta(SEXP sSEXP, SEXP d_SEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP R_SEXP, SEXP R_dot_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP LambdaSEXP, SEXP lambdaSEXP, SEXP betaSEXP, SEXP thetaSEXP, SEXP frailtySEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -501,7 +501,7 @@ END_RCPP
 }
 // dH_dtheta
 List dH_dtheta(NumericVector d_, List X_, List K_, List R_, List R_dot_, List R_star, List phi_1_, List phi_2_, List phi_3_, List phi_prime_1_, List phi_prime_2_, NumericVector Lambda, NumericVector lambda, NumericVector beta);
-RcppExport SEXP frailtyr_dH_dtheta(SEXP d_SEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP R_SEXP, SEXP R_dot_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP LambdaSEXP, SEXP lambdaSEXP, SEXP betaSEXP) {
+RcppExport SEXP frailtySurv_dH_dtheta(SEXP d_SEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP R_SEXP, SEXP R_dot_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP LambdaSEXP, SEXP lambdaSEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -525,7 +525,7 @@ END_RCPP
 }
 // jacobian_beta_beta
 double jacobian_beta_beta(int l, List X_, List K_, List H_, List phi_1_, List phi_2_, List phi_3_, List dH_dbeta_, List dH_dot_dbeta_);
-RcppExport SEXP frailtyr_jacobian_beta_beta(SEXP lSEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP dH_dbeta_SEXP, SEXP dH_dot_dbeta_SEXP) {
+RcppExport SEXP frailtySurv_jacobian_beta_beta(SEXP lSEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP dH_dbeta_SEXP, SEXP dH_dot_dbeta_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -544,7 +544,7 @@ END_RCPP
 }
 // jacobian_beta_theta
 double jacobian_beta_theta(int l, List X_, List K_, List H_, List phi_1_, List phi_2_, List phi_3_, List phi_prime_1_, List phi_prime_2_, List dH_dtheta_, List dH_dot_dtheta_);
-RcppExport SEXP frailtyr_jacobian_beta_theta(SEXP lSEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP dH_dtheta_SEXP, SEXP dH_dot_dtheta_SEXP) {
+RcppExport SEXP frailtySurv_jacobian_beta_theta(SEXP lSEXP, SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP dH_dtheta_SEXP, SEXP dH_dot_dtheta_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -565,7 +565,7 @@ END_RCPP
 }
 // jacobian_theta_beta
 double jacobian_theta_beta(List phi_1_, List phi_2_, List phi_prime_1_, List phi_prime_2_, List dH_dot_dbeta_);
-RcppExport SEXP frailtyr_jacobian_theta_beta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP dH_dot_dbeta_SEXP) {
+RcppExport SEXP frailtySurv_jacobian_theta_beta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP dH_dot_dbeta_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -580,7 +580,7 @@ END_RCPP
 }
 // jacobian_theta_theta
 double jacobian_theta_theta(List phi_1_, List phi_2_, List phi_prime_1_, List phi_prime_2_, List phi_prime_prime_1_, List dH_dot_dtheta_);
-RcppExport SEXP frailtyr_jacobian_theta_theta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP phi_prime_prime_1_SEXP, SEXP dH_dot_dtheta_SEXP) {
+RcppExport SEXP frailtySurv_jacobian_theta_theta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP, SEXP phi_prime_prime_1_SEXP, SEXP dH_dot_dtheta_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -596,7 +596,7 @@ END_RCPP
 }
 // Q_beta
 List Q_beta(List X_, List K_, List H_, List R_star, List phi_1_, List phi_2_, List phi_3_, int r);
-RcppExport SEXP frailtyr_Q_beta(SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP rSEXP) {
+RcppExport SEXP frailtySurv_Q_beta(SEXP X_SEXP, SEXP K_SEXP, SEXP H_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP, SEXP rSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -614,7 +614,7 @@ END_RCPP
 }
 // Q_theta
 List Q_theta(List H_, List R_star, List phi_1_, List phi_2_, List phi_prime_1_, List phi_prime_2_);
-RcppExport SEXP frailtyr_Q_theta(SEXP H_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP) {
+RcppExport SEXP frailtySurv_Q_theta(SEXP H_SEXP, SEXP R_starSEXP, SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_prime_1_SEXP, SEXP phi_prime_2_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -630,7 +630,7 @@ END_RCPP
 }
 // Ycal
 NumericVector Ycal(List X_, List R_star, List Y_, List psi_, NumericVector beta);
-RcppExport SEXP frailtyr_Ycal(SEXP X_SEXP, SEXP R_starSEXP, SEXP Y_SEXP, SEXP psi_SEXP, SEXP betaSEXP) {
+RcppExport SEXP frailtySurv_Ycal(SEXP X_SEXP, SEXP R_starSEXP, SEXP Y_SEXP, SEXP psi_SEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -645,7 +645,7 @@ END_RCPP
 }
 // eta
 List eta(List phi_1_, List phi_2_, List phi_3_);
-RcppExport SEXP frailtyr_eta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP) {
+RcppExport SEXP frailtySurv_eta(SEXP phi_1_SEXP, SEXP phi_2_SEXP, SEXP phi_3_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -658,7 +658,7 @@ END_RCPP
 }
 // Upsilon
 NumericVector Upsilon(List X_, List R_star, List K_, List R_dot_, List eta_, NumericVector Ycal_, NumericVector beta);
-RcppExport SEXP frailtyr_Upsilon(SEXP X_SEXP, SEXP R_starSEXP, SEXP K_SEXP, SEXP R_dot_SEXP, SEXP eta_SEXP, SEXP Ycal_SEXP, SEXP betaSEXP) {
+RcppExport SEXP frailtySurv_Upsilon(SEXP X_SEXP, SEXP R_starSEXP, SEXP K_SEXP, SEXP R_dot_SEXP, SEXP eta_SEXP, SEXP Ycal_SEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -675,7 +675,7 @@ END_RCPP
 }
 // Omega
 List Omega(List X_, List R_star, List N_, List R_dot_, List eta_, NumericVector Ycal_, NumericVector beta);
-RcppExport SEXP frailtyr_Omega(SEXP X_SEXP, SEXP R_starSEXP, SEXP N_SEXP, SEXP R_dot_SEXP, SEXP eta_SEXP, SEXP Ycal_SEXP, SEXP betaSEXP) {
+RcppExport SEXP frailtySurv_Omega(SEXP X_SEXP, SEXP R_starSEXP, SEXP N_SEXP, SEXP R_dot_SEXP, SEXP eta_SEXP, SEXP Ycal_SEXP, SEXP betaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -692,7 +692,7 @@ END_RCPP
 }
 // p_hat
 NumericVector p_hat(List I_, NumericVector Upsilon_, List Omega_, List N_tilde_);
-RcppExport SEXP frailtyr_p_hat(SEXP I_SEXP, SEXP Upsilon_SEXP, SEXP Omega_SEXP, SEXP N_tilde_SEXP) {
+RcppExport SEXP frailtySurv_p_hat(SEXP I_SEXP, SEXP Upsilon_SEXP, SEXP Omega_SEXP, SEXP N_tilde_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -706,7 +706,7 @@ END_RCPP
 }
 // pi_r
 NumericVector pi_r(List Q_, List N_tilde_, NumericVector p_hat);
-RcppExport SEXP frailtyr_pi_r(SEXP Q_SEXP, SEXP N_tilde_SEXP, SEXP p_hatSEXP) {
+RcppExport SEXP frailtySurv_pi_r(SEXP Q_SEXP, SEXP N_tilde_SEXP, SEXP p_hatSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -719,7 +719,7 @@ END_RCPP
 }
 // G_rl
 double G_rl(NumericVector pi_r, NumericVector pi_l, NumericVector p_hat, NumericVector Ycal_, List N_);
-RcppExport SEXP frailtyr_G_rl(SEXP pi_rSEXP, SEXP pi_lSEXP, SEXP p_hatSEXP, SEXP Ycal_SEXP, SEXP N_SEXP) {
+RcppExport SEXP frailtySurv_G_rl(SEXP pi_rSEXP, SEXP pi_lSEXP, SEXP p_hatSEXP, SEXP Ycal_SEXP, SEXP N_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -734,7 +734,7 @@ END_RCPP
 }
 // M_hat
 List M_hat(List X_, List R_star, List N_, List Y_, List psi_, NumericVector beta, NumericVector Lambda);
-RcppExport SEXP frailtyr_M_hat(SEXP X_SEXP, SEXP R_starSEXP, SEXP N_SEXP, SEXP Y_SEXP, SEXP psi_SEXP, SEXP betaSEXP, SEXP LambdaSEXP) {
+RcppExport SEXP frailtySurv_M_hat(SEXP X_SEXP, SEXP R_starSEXP, SEXP N_SEXP, SEXP Y_SEXP, SEXP psi_SEXP, SEXP betaSEXP, SEXP LambdaSEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;
@@ -751,7 +751,7 @@ END_RCPP
 }
 // u_star
 NumericMatrix u_star(List pi_, NumericVector p_hat, NumericVector Ycal_, List M_hat_);
-RcppExport SEXP frailtyr_u_star(SEXP pi_SEXP, SEXP p_hatSEXP, SEXP Ycal_SEXP, SEXP M_hat_SEXP) {
+RcppExport SEXP frailtySurv_u_star(SEXP pi_SEXP, SEXP p_hatSEXP, SEXP Ycal_SEXP, SEXP M_hat_SEXP) {
 BEGIN_RCPP
     Rcpp::RObject __result;
     Rcpp::RNGScope __rngScope;